+++ netstandard/Spring.Rest/Http/Client/Interceptor/IClientHttpRequestSyncExecution.cs
@@ -1,4 +1,4 @@
-﻿#if !SILVERLIGHT
+﻿
 #region License
 
 /*
@@ -44,4 +44,3 @@ namespace Spring.Http.Client.Interceptor
         IClientHttpResponse Execute();
     }
 }
-#endif
