+++ netstandard/Spring.Rest/Http/Client/Interceptor/InterceptingClientHttpRequest.cs
@@ -100,8 +100,6 @@ namespace Spring.Http.Client.Interceptor
                 this.targetRequest.Body = value;
             }
         }
-
-#if !SILVERLIGHT
         /// <summary>
         /// Execute this request, resulting in a <see cref="IClientHttpResponse" /> that can be read.
         /// </summary>
@@ -111,9 +109,6 @@ namespace Spring.Http.Client.Interceptor
             RequestSyncExecution requestSyncExecution = new RequestSyncExecution(this.targetRequest, this.body, this.interceptors);
             return requestSyncExecution.Execute();
         }
-#endif
-
-#if !CF_3_5
         /// <summary>
         /// Execute this request asynchronously.
         /// </summary>
@@ -137,7 +132,6 @@ namespace Spring.Http.Client.Interceptor
         {
             this.targetRequest.CancelAsync();
         }
-#endif
 
         #endregion
 
@@ -181,7 +175,6 @@ namespace Spring.Http.Client.Interceptor
             }
         }
 
-#if !SILVERLIGHT
         private sealed class RequestSyncExecution : AbstractRequestContext, IClientHttpRequestSyncExecution
         {
             public RequestSyncExecution(
@@ -212,9 +205,7 @@ namespace Spring.Http.Client.Interceptor
                 }
             }
         }
-#endif
 
-#if !CF_3_5
         private sealed class RequestAsyncExecution : AbstractRequestContext, IClientHttpRequestAsyncExecution
         {
             private object asyncState;
@@ -362,7 +353,6 @@ namespace Spring.Http.Client.Interceptor
                 }
             }
         }
-#endif
 
         #endregion
     }
