+++ netstandard/Spring.Rest/Http/Converters/HttpMessageConversionException.cs
@@ -28,9 +28,7 @@ namespace Spring.Http.Converters
     /// </summary>
     /// <author>Arjen Poutsma</author>
     /// <author>Bruno Baia (.NET)</author>
-#if !SILVERLIGHT && !CF_3_5
     [Serializable]
-#endif
     public class HttpMessageConversionException : Exception
     {
         /// <summary>
@@ -65,7 +63,6 @@ namespace Spring.Http.Converters
         {
         }
 
-#if !SILVERLIGHT && !CF_3_5
         /// <summary>
         /// Creates a new instance of the <see cref="HttpMessageConversionException"/> class.
         /// </summary>
@@ -81,6 +78,5 @@ namespace Spring.Http.Converters
             : base(info, context)
         {
         }
-#endif
     }
 }
