+++ netstandard/Spring.Rest/Http/Converters/StringHttpMessageConverter.cs
@@ -39,22 +39,11 @@ namespace Spring.Http.Converters
         /// <summary>
         /// Default encoding for strings.
         /// </summary>
-#if SILVERLIGHT || CF_3_5
-        protected static readonly Encoding DEFAULT_CHARSET = new UTF8Encoding(false); // Remove byte Order Mask (BOM)
-#else
-        protected static readonly Encoding DEFAULT_CHARSET = Encoding.GetEncoding("ISO-8859-1");
-#endif
 
-#if SILVERLIGHT || CF_3_5
-        /// <summary>
-        /// Creates a new instance of the <see cref="StringHttpMessageConverter"/> 
-        /// with 'text/plain; charset=utf-8', and '*/*' media types.
-        /// </summary>
-        public StringHttpMessageConverter() :
-            base(new MediaType("text", "plain", "UTF-8"), MediaType.ALL)
-        {
-        }
-#else
+        protected static readonly Encoding DEFAULT_CHARSET = Encoding.GetEncoding("ISO-8859-1");
+
+
+
         /// <summary>
         /// Creates a new instance of the <see cref="StringHttpMessageConverter"/> 
         /// with 'text/plain; charset=ISO-8859-1', and '*/*' media types.
@@ -63,7 +52,7 @@ namespace Spring.Http.Converters
             base(new MediaType("text", "plain", "ISO-8859-1"), MediaType.ALL)
         {
         }
-#endif
+
 
         /// <summary>
         /// Indicates whether the given class is supported by this converter.
@@ -107,11 +96,6 @@ namespace Spring.Http.Converters
 
             // Create a byte array of the data we want to send  
             byte[] byteData = encoding.GetBytes(content as string);
-
-//#if !SILVERLIGHT
-//            // Set the content length in the message headers  
-//            message.Headers.ContentLength = byteData.Length;
-//#endif
 
             // Write to the message stream
             message.Body = delegate(Stream stream)
