+++ netstandard/Spring.Rest/Http/Converters/Xml/XElementHttpMessageConverter.cs
@@ -1,4 +1,4 @@
-﻿#if NET_3_5 || SILVERLIGHT_XML_LINQ || WINDOWS_PHONE || CF_3_5
+﻿
 #region License
 
 /*
@@ -78,4 +78,3 @@ namespace Spring.Http.Converters.Xml
         }
     }
 }
-#endif
