+++ netstandard/Spring.Rest/Http/HttpResponseMessage`1.cs
@@ -29,9 +29,7 @@ namespace Spring.Http
     /// </summary>
     /// <typeparam name="T">The type of the response body.</typeparam>
     /// <author>Bruno Baia</author>
-#if !SILVERLIGHT
     [Serializable]
-#endif
     public class HttpResponseMessage<T> : HttpResponseMessage where T : class
     {
         private T body;
