+++ netstandard/Spring.Rest/Rest/Client/RestClientException.cs
@@ -28,9 +28,7 @@ namespace Spring.Rest.Client
     /// </summary>
     /// <author>Arjen Poutsma</author>
     /// <author>Bruno Baia (.NET)</author>
-#if !SILVERLIGHT && !CF_3_5
     [Serializable]
-#endif
     public class RestClientException : Exception
     {
         /// <summary>
@@ -65,7 +63,6 @@ namespace Spring.Rest.Client
         {
         }
 
-#if !SILVERLIGHT && !CF_3_5
         /// <summary>
         /// Creates a new instance of the <see cref="RestClientException"/> class.
         /// </summary>
@@ -81,6 +78,5 @@ namespace Spring.Rest.Client
             : base(info, context)
         {
         }
-#endif
     }
 }
