+++ netstandard/Spring.Rest/Rest/Client/RestOperationCanceler.cs
@@ -1,4 +1,4 @@
-﻿#if !CF_3_5
+﻿
 #region License
 
 /*
@@ -80,4 +80,3 @@ namespace Spring.Rest.Client
         }
     }
 }
-#endif
