+++ netstandard/Spring.Rest/Rest/Client/RestTemplate.cs
@@ -20,22 +20,17 @@
 
 using System;
 using System.Collections.Generic;
-#if NET_4_0 || SILVERLIGHT_5 || NETSTANDARD || NETFRAMEWORK
 using System.Threading;
 using System.Threading.Tasks;
-#endif
 
 using Spring.Http;
 using Spring.Http.Client;
 using Spring.Http.Client.Interceptor;
 using Spring.Http.Converters;
 using Spring.Http.Converters.Xml;
-#if NET_3_5 || SILVERLIGHT
+
 using Spring.Http.Converters.Json;
-#endif
-#if NET_3_5 && !SILVERLIGHT
-using Spring.Http.Converters.Feed;
-#endif
+
 using Spring.Rest.Client.Support;
 using UriTemplate = Spring.Util.UriTemplate; // UriTemplate exists in .NET Framework since 3.5
 using ArgumentUtils = Spring.Util.ArgumentUtils;
@@ -111,7 +106,7 @@ namespace Spring.Rest.Client
     /// <see cref="IResponseExtractor{T}"/>
     /// <author>Arjen Poutsma</author>
     /// <author>Bruno Baia (.NET)</author>
-    public class RestTemplate : IRestOperations
+    public partial class RestTemplate : IRestOperations
     {
         #region Logging
 #if !SILVERLIGHT && !CF_3_5
@@ -228,16 +223,7 @@ namespace Spring.Rest.Client
             this._messageConverters.Add(new ByteArrayHttpMessageConverter());
             this._messageConverters.Add(new StringHttpMessageConverter());
             this._messageConverters.Add(new FormHttpMessageConverter());
-#if !SILVERLIGHT
             this._messageConverters.Add(new XmlDocumentHttpMessageConverter());
-#endif
-#if NET_3_5 || WINDOWS_PHONE
-            this._messageConverters.Add(new XElementHttpMessageConverter());
-#if NET_3_5
-            this._messageConverters.Add(new Rss20FeedHttpMessageConverter());
-            this._messageConverters.Add(new Atom10FeedHttpMessageConverter());
-#endif
-#endif
         }
 
         #endregion
@@ -246,7 +232,6 @@ namespace Spring.Rest.Client
 
         #region Synchronous operations
 
-#if !SILVERLIGHT
         #region GET
 
         /// <summary>
@@ -970,947 +955,16 @@ namespace Spring.Rest.Client
         }
 
         #endregion
-#endif
+
 
         #endregion
+        
 
-        #region Asynchronous operations
-
-#if !CF_3_5
-        #region GET
-
-        /// <summary>
-        /// Asynchronously retrieve a representation by doing a GET on the specified URL. 
-        /// The response (if any) is converted.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForObjectAsync<T>(string url, Action<RestOperationCompletedEventArgs<T>> getCompleted, params object[] uriVariables) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.GET, requestCallback, responseExtractor, getCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve a representation by doing a GET on the specified URL. 
-        /// The response (if any) is converted.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForObjectAsync<T>(string url, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<T>> getCompleted) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.GET, requestCallback, responseExtractor, uriVariables, getCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve a representation by doing a GET on the specified URL. 
-        /// The response (if any) is converted.
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForObjectAsync<T>(Uri url, Action<RestOperationCompletedEventArgs<T>> getCompleted) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.GET, requestCallback, responseExtractor, getCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve an entity by doing a GET on the specified URL. 
-        /// The response is converted and stored in an <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForMessageAsync<T>(string url, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> getCompleted, params object[] uriVariables) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.GET, requestCallback, responseExtractor, getCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve an entity by doing a GET on the specified URL. 
-        /// The response is converted and stored in an <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForMessageAsync<T>(string url, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> getCompleted) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.GET, requestCallback, responseExtractor, uriVariables, getCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve an entity by doing a GET on the specified URL. 
-        /// The response is converted and stored in an <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="getCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous GET method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler GetForMessageAsync<T>(Uri url, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> getCompleted) where T : class
-        {
-            AcceptHeaderRequestCallback requestCallback = new AcceptHeaderRequestCallback(typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.GET, requestCallback, responseExtractor, getCompleted);
-        }
-
-        #endregion
-
-        #region HEAD
-
-        /// <summary>
-        /// Asynchronously retrieve all headers of the resource specified by the URI template.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="headCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous HEAD method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler HeadForHeadersAsync(string url, Action<RestOperationCompletedEventArgs<HttpHeaders>> headCompleted, params object[] uriVariables)
-        {
-            HeadersResponseExtractor responseExtractor = new HeadersResponseExtractor();
-            return this.ExecuteAsync<HttpHeaders>(url, HttpMethod.HEAD, null, responseExtractor, headCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve all headers of the resource specified by the URI template.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="headCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous HEAD method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler HeadForHeadersAsync(string url, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpHeaders>> headCompleted)
-        {
-            HeadersResponseExtractor responseExtractor = new HeadersResponseExtractor();
-            return this.ExecuteAsync<HttpHeaders>(url, HttpMethod.HEAD, null, responseExtractor, uriVariables, headCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously retrieve all headers of the resource specified by the URI template.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="headCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous HEAD method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler HeadForHeadersAsync(Uri url, Action<RestOperationCompletedEventArgs<HttpHeaders>> headCompleted)
-        {
-            HeadersResponseExtractor responseExtractor = new HeadersResponseExtractor();
-            return this.ExecuteAsync<HttpHeaders>(url, HttpMethod.HEAD, null, responseExtractor, headCompleted);
-        }
-
-        #endregion
-
-        #region POST
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the value of the 'Location' header. 
-        /// This header typically indicates where the new resource is stored.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForLocationAsync(string url, object request, Action<RestOperationCompletedEventArgs<Uri>> postCompleted, params object[] uriVariables)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            LocationHeaderResponseExtractor responseExtractor = new LocationHeaderResponseExtractor();
-            return this.ExecuteAsync<Uri>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the value of the 'Location' header. 
-        /// This header typically indicates where the new resource is stored.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForLocationAsync(string url, object request, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<Uri>> postCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            LocationHeaderResponseExtractor responseExtractor = new LocationHeaderResponseExtractor();
-            return this.ExecuteAsync<Uri>(url, HttpMethod.POST, requestCallback, responseExtractor, uriVariables, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the value of the 'Location' header. 
-        /// This header typically indicates where the new resource is stored.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForLocationAsync(Uri url, object request, Action<RestOperationCompletedEventArgs<Uri>> postCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            LocationHeaderResponseExtractor responseExtractor = new LocationHeaderResponseExtractor();
-            return this.ExecuteAsync<Uri>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the representation found in the response. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForObjectAsync<T>(string url, object request, Action<RestOperationCompletedEventArgs<T>> postCompleted, params object[] uriVariables) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the representation found in the response. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForObjectAsync<T>(string url, object request, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<T>> postCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.POST, requestCallback, responseExtractor, uriVariables, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the representation found in the response. 
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForObjectAsync<T>(Uri url, object request, Action<RestOperationCompletedEventArgs<T>> postCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            MessageConverterResponseExtractor<T> responseExtractor = new MessageConverterResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<T>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync<T>(string url, object request, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> postCompleted, params object[] uriVariables) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync<T>(string url, object request, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> postCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.POST, requestCallback, responseExtractor, uriVariables, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>. 
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync<T>(Uri url, object request, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> postCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync(string url, object request, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> postCompleted, params object[] uriVariables)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>. 
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync(string url, object request, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> postCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, HttpMethod.POST, requestCallback, responseExtractor, uriVariables, postCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create a new resource by POSTing the given object to the URI template, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>. 
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="postCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous POST method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PostForMessageAsync(Uri url, object request, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> postCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, HttpMethod.POST, requestCallback, responseExtractor, postCompleted);
-        }
-
-        #endregion
-
-        #region PUT
-
-        /// <summary>
-        /// Asynchronously create or update a resource by PUTting the given object to the URI.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="putCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PutAsync(string url, object request, Action<RestOperationCompletedEventArgs<object>> putCompleted, params object[] uriVariables)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            return this.ExecuteAsync(url, HttpMethod.PUT, requestCallback, null, putCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously create or update a resource by PUTting the given object to the URI.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </para>
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="putCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PutAsync(string url, object request, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<object>> putCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            return this.ExecuteAsync(url, HttpMethod.PUT, requestCallback, null, uriVariables, putCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously create or update a resource by PUTting the given object to the URI.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="request">
-        /// The object to be POSTed, may be a <see cref="HttpEntity"/> in order to add additional HTTP headers.
-        /// </param>
-        /// <param name="putCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler PutAsync(Uri url, object request, Action<RestOperationCompletedEventArgs<object>> putCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(request, this._messageConverters);
-            return this.ExecuteAsync(url, HttpMethod.PUT, requestCallback, null, putCompleted);
-        }
-
-        #endregion
-
-        #region DELETE
-
-        /// <summary>
-        /// Asynchronously delete the resources at the specified URI.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="deleteCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler DeleteAsync(string url, Action<RestOperationCompletedEventArgs<object>> deleteCompleted, params object[] uriVariables)
-        {
-            return this.ExecuteAsync(url, HttpMethod.DELETE, null, null, deleteCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously delete the resources at the specified URI.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="deleteCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler DeleteAsync(string url, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<object>> deleteCompleted)
-        {
-            return this.ExecuteAsync(url, HttpMethod.DELETE, null, null, uriVariables, deleteCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously delete the resources at the specified URI.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="deleteCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous PUT method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler DeleteAsync(Uri url, Action<RestOperationCompletedEventArgs<object>> deleteCompleted)
-        {
-            return this.ExecuteAsync(url, HttpMethod.DELETE, null, null, deleteCompleted);
-        }
-
-        #endregion
-
-        #region OPTIONS
-
-        /// <summary>
-        /// Asynchronously return the value of the Allow header for the given URI.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="optionsCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous OPTIONS method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler OptionsForAllowAsync(string url, Action<RestOperationCompletedEventArgs<IList<HttpMethod>>> optionsCompleted, params object[] uriVariables)
-        {
-            AllowHeaderResponseExtractor responseExtractor = new AllowHeaderResponseExtractor();
-            return this.ExecuteAsync<IList<HttpMethod>>(url, HttpMethod.OPTIONS, null, responseExtractor, optionsCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously return the value of the Allow header for the given URI.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="optionsCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous OPTIONS method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler OptionsForAllowAsync(string url, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<IList<HttpMethod>>> optionsCompleted)
-        {
-            AllowHeaderResponseExtractor responseExtractor = new AllowHeaderResponseExtractor();
-            return this.ExecuteAsync<IList<HttpMethod>>(url, HttpMethod.OPTIONS, null, responseExtractor, uriVariables, optionsCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously return the value of the Allow header for the given URI.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="optionsCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous OPTIONS method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler OptionsForAllowAsync(Uri url, Action<RestOperationCompletedEventArgs<IList<HttpMethod>>> optionsCompleted)
-        {
-            AllowHeaderResponseExtractor responseExtractor = new AllowHeaderResponseExtractor();
-            return this.ExecuteAsync<IList<HttpMethod>>(url, HttpMethod.OPTIONS, null, responseExtractor, optionsCompleted);
-        }
-
-        #endregion
-
-
-        #region Exchange
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync<T>(string url, HttpMethod method, HttpEntity requestEntity, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> methodCompleted, params object[] uriVariables) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, method, requestCallback, responseExtractor, methodCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync<T>(string url, HttpMethod method, HttpEntity requestEntity, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> methodCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, method, requestCallback, responseExtractor, uriVariables, methodCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response as <see cref="HttpResponseMessage{T}"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync<T>(Uri url, HttpMethod method, HttpEntity requestEntity, Action<RestOperationCompletedEventArgs<HttpResponseMessage<T>>> methodCompleted) where T : class
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, typeof(T), this._messageConverters);
-            HttpMessageResponseExtractor<T> responseExtractor = new HttpMessageResponseExtractor<T>(this._messageConverters);
-            return this.ExecuteAsync<HttpResponseMessage<T>>(url, method, requestCallback, responseExtractor, methodCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync(string url, HttpMethod method, HttpEntity requestEntity, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> methodCompleted, params object[] uriVariables)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, method, requestCallback, responseExtractor, methodCompleted, uriVariables);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync(string url, HttpMethod method, HttpEntity requestEntity, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> methodCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, method, requestCallback, responseExtractor, uriVariables, methodCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, writing the given request message to the request, 
-        /// and returns the response with no entity as <see cref="HttpResponseMessage"/>.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestEntity">
-        /// The HTTP entity (headers and/or body) to write to the request, may be <see langword="null"/>.
-        /// </param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExchangeAsync(Uri url, HttpMethod method, HttpEntity requestEntity, Action<RestOperationCompletedEventArgs<HttpResponseMessage>> methodCompleted)
-        {
-            HttpEntityRequestCallback requestCallback = new HttpEntityRequestCallback(requestEntity, this._messageConverters);
-            HttpMessageResponseExtractor responseExtractor = new HttpMessageResponseExtractor();
-            return this.ExecuteAsync<HttpResponseMessage>(url, method, requestCallback, responseExtractor, methodCompleted);
-        }
-
-        #endregion
-
-        #region General execution
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, preparing the request with the 
-        /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given URI variables, if any.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestCallback">Object that prepares the request.</param>
-        /// <param name="responseExtractor">Object that extracts the return value from the response.</param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>   
-        /// <param name="uriVariables">The variables to expand the template.</param>
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExecuteAsync<T>(string url, HttpMethod method, IRequestCallback requestCallback, IResponseExtractor<T> responseExtractor, Action<RestOperationCompletedEventArgs<T>> methodCompleted, params object[] uriVariables) where T : class
-        {
-            return this.DoExecuteAsync<T>(BuildUri(this._baseAddress, url, uriVariables), method, 
-                requestCallback, responseExtractor, methodCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, preparing the request with the 
-        /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
-        /// </summary>
-        /// <remarks>
-        /// URI Template variables are expanded using the given dictionary.
-        /// </remarks>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestCallback">Object that prepares the request.</param>
-        /// <param name="responseExtractor">Object that extracts the return value from the response.</param>
-        /// <param name="uriVariables">The dictionary containing variables for the URI template.</param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>  
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExecuteAsync<T>(string url, HttpMethod method, IRequestCallback requestCallback, IResponseExtractor<T> responseExtractor, IDictionary<string, object> uriVariables, Action<RestOperationCompletedEventArgs<T>> methodCompleted) where T : class
-        {
-            return this.DoExecuteAsync<T>(BuildUri(this._baseAddress, url, uriVariables), method, 
-                requestCallback, responseExtractor, methodCompleted);
-        }
-
-        /// <summary>
-        /// Asynchronously execute the HTTP method to the given URI template, preparing the request with the 
-        /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
-        /// </summary>
-        /// <typeparam name="T">The type of the response value.</typeparam>
-        /// <param name="url">The URL.</param>
-        /// <param name="method">The HTTP method (GET, POST, etc.)</param>
-        /// <param name="requestCallback">Object that prepares the request.</param>
-        /// <param name="responseExtractor">Object that extracts the return value from the response.</param>
-        /// <param name="methodCompleted">
-        /// The <code>Action&lt;T&gt;</code> to perform when the asynchronous method completes.
-        /// </param>  
-        /// <returns>
-        /// A <see cref="RestOperationCanceler"/> instance that allows to cancel the asynchronous operation.
-        /// </returns>
-        public RestOperationCanceler ExecuteAsync<T>(Uri url, HttpMethod method, IRequestCallback requestCallback, IResponseExtractor<T> responseExtractor, Action<RestOperationCompletedEventArgs<T>> methodCompleted) where T : class
-        {
-            return this.DoExecuteAsync<T>(BuildUri(this._baseAddress, url), method, 
-                requestCallback, responseExtractor, methodCompleted);
-        }
-
-        #endregion
-#endif
-
-        #endregion
+        
 
         #region Asynchronous operations (TPL)
 
-#if     SILVERLIGHT_5 || NETSTANDARD || NETFRAMEWORK
+
         #region GET
 
         /// <summary>
@@ -2654,7 +1708,6 @@ namespace Spring.Rest.Client
         }
 
         #endregion
-#endif
 
         #endregion
 
@@ -2662,7 +1715,6 @@ namespace Spring.Rest.Client
 
         #region DoExecute
 
-#if !SILVERLIGHT
         /// <summary>
         /// Execute the HTTP request to the given URI, preparing the request with the 
         /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
@@ -2704,13 +1756,11 @@ namespace Spring.Rest.Client
 
             return null;
         }
-#endif
 
         #endregion
 
         #region DoExecuteAsync
 
-#if !CF_3_5
         /// <summary>
         /// Asynchronously execute the HTTP request to the given URI, preparing the request with the 
         /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
@@ -2798,13 +1848,11 @@ namespace Spring.Rest.Client
                 state.MethodCompleted(new RestOperationCompletedEventArgs<T>(value, exception, cancelled, null));
             }
         }
-#endif
 
         #endregion
 
         #region DoExecuteAsync (TPL)
 
-#if NETFRAMEWORK || NETSTANDARD || SILVERLIGHT_5 
         /// <summary>
         /// Asynchronously execute the HTTP request to the given URI, preparing the request with the 
         /// <see cref="IRequestCallback"/>, and reading the response with an <see cref="IResponseExtractor{T}"/>.
@@ -2849,7 +1897,6 @@ namespace Spring.Rest.Client
 
             return taskCompletionSource.Task;
         }
-#endif
 
         #endregion
 
@@ -2930,14 +1977,12 @@ namespace Spring.Rest.Client
         private static void LogResponseStatus(Uri uri, HttpMethod method, IClientHttpResponse response) 
         {
             #region Instrumentation
-#if !SILVERLIGHT && !CF_3_5
             if (LOG.IsDebugEnabled)
             {
-                LOG.Debug(String.Format(
+                LOG.Debug(string.Format(
                     "{0} request for '{1}' resulted in {2:d} - {2} ({3})",
                     method, uri, response.StatusCode, response.StatusDescription));
             }
-#endif
             #endregion
         }
 
@@ -2945,14 +1990,12 @@ namespace Spring.Rest.Client
             IResponseErrorHandler errorHandler)
         {
             #region Instrumentation
-#if !SILVERLIGHT && !CF_3_5
             if (LOG.IsWarnEnabled)
             {
                 LOG.Warn(String.Format(
                     "{0} request for '{1}' resulted in {2:d} - {2} ({3}); invoking error handler",
                     method, uri, response.StatusCode, response.StatusDescription));
             }
-#endif
             #endregion
 
             errorHandler.HandleError(uri, method, response);
