+++ netstandard/Spring.Rest/Rest/Client/Support/RestAsyncOperationState.cs
@@ -1,4 +1,4 @@
-﻿#if !CF_3_5
+﻿
 #region License
 
 /*
@@ -114,4 +114,3 @@ namespace Spring.Rest.Client.Support
         }
     }
 }
-#endif
